Refactor code for better modularity
