+++ index.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="main.css">
     <!-- Load d3.js and main script-->
     <script src="https://d3js.org/d3.v6.js"></script>
-    <script src="main.js" defer></script>
+    <script type="module" src="main.js" defer></script>
 </head>
 <body>
     <!-- Home page section -->
